Ho fatto un repulisti ... ho aggiunto la visualizzazione di Elisa e contestualmente la ho neutralizzata.

+++ trunk/elanja-fltk/gui_controls.cpp
@@ -8,7 +8,7 @@ int gui_agents;
 int gui_distance;
 double gui_rho;
 int gui_nFeatures;
-double gui_friends;
+int gui_friends;
 double gui_threshold;
 
 bool doNextSimulationStep;
